pyproject.toml
[project]
name = "mcp-t1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "openai-agents>=0.14.6",
    "python-dotenv>=1.2.2",
]
